[posix] E: Move stubs to per-header libc_* crates
Move the remaining C-ABI stub functions out of posix/src/dummy.rs and
posix/src/grp/ into dedicated per-header libc_* crates, matching the
existing split pattern where posix re-exports them via `extern crate`:

  - libc_grp:         getgrgid getgrnam getgrouplist setgrent endgrent
                      getgrent initgroups
  - libc_termios:     tcgetattr tcsetattr
  - libc_unistd:      pathconf fpathconf
  - libc_ftw:         ftw
  - libc_sys_statvfs: statvfs fstatvfs
  - libc_dirent:      fdopendir
  - libc_glob:        glob globfree (+ GlobT, GLOB_* consts)
  - libc_stdio:       popen pclose

Also declare fdopendir in dirent.h via its generated spec
(sysapi/headers/dirent.toml) for POSIX conformance.
