+++ src/libs/libc_dirent/src/lib.rs
@@ -2,8 +2,13 @@
 // Licensed under the MIT License.
 
 //==================================================================================================
+// Crate Configuration
+//==================================================================================================
+
+#![cfg_attr(not(feature = "std"), no_std)]
+
+//==================================================================================================
 // Modules
 //==================================================================================================
 
-#[cfg(feature = "syscall")]
 mod bindings;
