+++ src/libs/libc_stdio/src/lib.rs
@@ -69,7 +69,9 @@ pub mod ftello;
 pub mod fwrite;
 pub mod getchar;
 pub mod getdelim;
+pub mod pclose;
 pub mod perror;
+pub mod popen;
 pub mod printf;
 pub mod putchar;
 pub mod puts;
