+++ src/libs/libc_termios/src/lib.rs
@@ -1,0 +1,14 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Crate Configuration
+//==================================================================================================
+
+#![cfg_attr(not(feature = "std"), no_std)]
+
+//==================================================================================================
+// Modules
+//==================================================================================================
+
+mod bindings;
