+++ src/libs/sysapi/headers/dirent.toml
@@ -79,9 +79,12 @@ struct posix_dent {
 
 [[sections]]
 title = "Directory Operations"
-functions = ["opendir", "readdir", "closedir"]
+functions = ["opendir", "fdopendir", "readdir", "closedir"]
 
 # opendir's Rust binding spells its path argument `*const i8`; declare the
 # standard POSIX prototype so C callers can pass a `const char *`.
+# fdopendir's Rust binding returns the opaque `*mut c_void` placeholder; declare
+# the standard POSIX prototype so C callers see the `DIR *` return type.
 [overrides]
 opendir = "extern DIR *opendir(const char *dirname);"
+fdopendir = "extern DIR *fdopendir(int fd);"
